pyproject.toml
[project]
name = "nomangho-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "httpx",
  "python-dotenv"
]
